remove trashy and redundunt tests + combined neat test for all flows
